rh9: import RHEL9 kernel-5.14.0-425.el9
https://access.redhat.com/downloads/content/kernel/5.14.0-425.el9/x86_64/fd431d51/package-changelog

https://virtuozzo.atlassian.net/browse/PSBM-154941

+++ kernel/trace/trace_syscalls.c
@@ -573,7 +573,7 @@ static int perf_call_bpf_enter(struct trace_event_call *call, struct pt_regs *re
 {
 	struct syscall_tp_t {
 		struct trace_entry ent;
-		unsigned long syscall_nr;
+		int syscall_nr;
 		unsigned long args[SYSCALL_DEFINE_MAXARGS];
 	} __aligned(8) param;
 	int i;
@@ -678,7 +678,7 @@ static int perf_call_bpf_exit(struct trace_event_call *call, struct pt_regs *reg
 {
 	struct syscall_tp_t {
 		struct trace_entry ent;
-		unsigned long syscall_nr;
+		int syscall_nr;
 		unsigned long ret;
 	} __aligned(8) param;
